+++ src/components/pages/Profile/ProfileEditForm.scss
@@ -1,0 +1,3 @@
+.profile__element {
+  margin: 20px;
+}
