+++ src/components/shared/ModalConfirm/Backdrop/Backdrop.scss
@@ -1,5 +1,3 @@
-@import "../../../assets/styles/colors.scss";
-
 .backdrop {
   position: absolute;
   top: 0;
